artifacts slot fix images

+++ src/app/infrastructure/persistence/models/player_orm.py
@@ -1,7 +1,7 @@
 from datetime import datetime
 from uuid import UUID
 from typing import List, TYPE_CHECKING
-from sqlalchemy import Integer, String, Date, DateTime, Uuid
+from sqlalchemy import BigInteger, Integer, String, Date, DateTime, Uuid
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from app.infrastructure.persistence.models.base import Base
@@ -14,7 +14,7 @@ class PlayerORM(Base):
     __tablename__ = "players"
 
     id: Mapped[UUID] = mapped_column(Uuid, primary_key=True, index=True)
-    telegram_id: Mapped[int] = mapped_column(Integer, index=True, unique=True)
+    telegram_id: Mapped[int] = mapped_column(BigInteger, index=True, unique=True)
     username: Mapped[str | None] = mapped_column(String(30), nullable=True)
     xp: Mapped[int] = mapped_column(Integer, default=0)
     xgen_balance: Mapped[int] = mapped_column(Integer, server_default="0")
